move ignore-keys to one level higher
