+++ packages/insomnia-app/app/sync/__tests__/ignore-keys.test.ts
@@ -1,5 +1,5 @@
 import { createBuilder } from '@develohpanda/fluent-builder';
-import { baseModelSchema, workspaceModelSchema } from '../../../models/__schemas__/model-schemas';
+import { baseModelSchema, workspaceModelSchema } from '../../models/__schemas__/model-schemas';
 import { deleteKeys, resetKeys, shouldIgnoreKey } from '../ignore-keys';
 
 const baseModelBuilder = createBuilder(baseModelSchema);
