+++ packages/insomnia-app/app/sync/ignore-keys.ts
@@ -1,5 +1,5 @@
-import { BaseModel } from '../../models';
-import { isWorkspace, Workspace } from '../../models/workspace';
+import { BaseModel } from '../models';
+import { isWorkspace, Workspace } from '../models/workspace';
 
 // Key for VCS to delete before computing changes
 const DELETE_KEY: keyof BaseModel = 'modified';
